fix: add Introduction step in the Single collection Edit view (#23943)
